Adapta order-service e passa a usar BigDecimal e UUID

+++ order-service/src/main/java/br/edu/atitus/order_service/clients/CurrencyClient.java
@@ -1,5 +1,7 @@
 package br.edu.atitus.order_service.clients;
 
+import java.math.BigDecimal;
+
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -9,7 +11,7 @@ public interface CurrencyClient {
 	
 	@GetMapping("/currency/{value}/{source}/{target}")
 	CurrencyResponse getCurrency(
-			@PathVariable double value,
+			@PathVariable BigDecimal value,
 			@PathVariable String source,
 			@PathVariable String target);
